feat(EDCO-07): cards component, added more complete stories
