+++ src/components/Cards/index.tsx
@@ -1,19 +1,9 @@
-import {
-  Card,
-  CardContent,
-  Grid,
-  Typography,
-  Stack,
-  Link,
-  Box,
-} from '@mui/material';
-import ArrowRightAltIcon from '@mui/icons-material/ArrowRightAlt';
-import { type CommonProps, type Generic } from '../../types/components';
-import EContainer from '../EContainer';
-import { EIcon, type EIconProps } from '../EIcon';
-
-import { isJSX } from '../../utils';
+import { Grid, Typography } from '@mui/material';
 import { Masonry } from '@mui/lab';
+import { type Generic, type CommonProps } from '../../types/components';
+import EContainer from '../EContainer';
+import Item, { type IItem } from './item';
+import { isJSX } from '../../utils';
 
 const layoutMap = new Map();
 
@@ -21,94 +11,13 @@ layoutMap.set('3-items', 3);
 layoutMap.set('4-items', 4);
 layoutMap.set('full-text', 2);
 
-interface IItem {
-  textAlign?: 'center' | 'left';
-  icon?: Generic | EIconProps['name'];
-  label?: string;
-  title: string;
-  text: string;
-  link?: {
-    href: string;
-    text: string;
-    title: string;
-  };
-}
-
-const Item = (props: IItem) => {
-  const { title, text, icon, link, textAlign, label } = props;
-  return (
-    <Card elevation={16}>
-      <CardContent>
-        <Stack
-          px={4}
-          sx={{ minHeight: '275px' }}
-          justifyContent="center"
-          alignItems={textAlign}
-        >
-          <Box mb={2} color="primary.dark">
-            {isJSX(icon) ? (
-              icon
-            ) : (
-              <EIcon name={icon} sx={{ fontSize: 40, color: 'inerith' }} />
-            )}
-          </Box>
-          {label && (
-            <Typography
-              mb={1}
-              textTransform="uppercase"
-              fontSize="12px"
-              fontWeight="600"
-              color="text.secondary"
-            >
-              {label}
-            </Typography>
-          )}
-          <Typography variant="h6" mb={1}>
-            {title}
-          </Typography>
-          <Typography
-            variant="body1"
-            mb={1}
-            textAlign={{ textAlign }}
-            fontSize="16px"
-          >
-            {text}
-          </Typography>
-          {link && (
-            <Stack
-              mt={2}
-              direction="row"
-              alignItems="center"
-              color="primary.main"
-              justifyContent={textAlign}
-            >
-              <Link
-                color="primary.main"
-                underline="none"
-                textTransform="capitalize"
-                href={link.href}
-                title={link.title}
-                fontSize={14}
-                fontWeight={400}
-              >
-                {link.text}
-              </Link>
-              <ArrowRightAltIcon sx={{ color: 'inerith', fontSize: 18 }} />
-            </Stack>
-          )}
-        </Stack>
-      </CardContent>
-    </Card>
-  );
-};
-
 export interface CardsProps extends CommonProps {
   items: IItem[];
   layout: '3-items' | '4-items' | 'full-text';
   text: {
     title: string;
     subtitle?: string;
-    body?: string;
+    body?: string | Generic;
   };
 }
 
@@ -127,9 +36,13 @@ const Cards = (props: CardsProps) => {
             <Typography variant="h6" mb={5} color={'inherit'}>
               {text.subtitle}
             </Typography>
-            <Typography variant="body1" color={'inherit'}>
-              {text.body}
-            </Typography>
+            {isJSX(text.body) ? (
+              text.body
+            ) : (
+              <Typography variant="body1" color={'inherit'}>
+                {text.body}
+              </Typography>
+            )}
           </Grid>
         )}
         <Grid item md={layout !== 'full-text' ? 12 : 8}>
